+++ src/main/java/br/com/promocoes_api/api/dto/promocao/PromocaoDTO.java
@@ -1,12 +1,12 @@
 package br.com.promocoes_api.api.dto.promocao;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 public record PromocaoDTO(
     Long id,
     String titulo,
-    BigDecimal desconto,
+    int desconto,
     LocalDate dataInicioPromo,
     LocalDate dataFimPromo
 ) {}
